+++ lib/index.ts
@@ -6,9 +6,9 @@ import { getUniResolver, getUniResolvers, UniResolver } from './resolver/UniReso
 import { Constants as DidUniConstants } from './types/constants';
 import { Config } from './types/types';
 import { DIDRegistrationResult } from './types/types';
-import { DIDDocument, DifResolver, VerificationMethod } from './types/types';
+import { DIDDocument, VerificationMethod } from './types/types';
 
-export { DifResolver, DIDDocument, VerificationMethod };
+export { DIDDocument, VerificationMethod };
 export { DIDResolutionResult };
 export { DIDRegistrationResult };